+++ src/main/java/cart/authentication/Authentication.java
@@ -1,0 +1,12 @@
+package cart.authentication;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.PARAMETER)
+public @interface Authentication {
+
+}
