+++ src/main/java/cart/authentication/MemberAuthenticationException.java
@@ -1,4 +1,4 @@
-package cart.util;
+package cart.authentication;
 
 public class MemberAuthenticationException extends RuntimeException {
     public MemberAuthenticationException(String message) {
